Update BARE_METAL.md (#4175)
Pull Request Type
[ ] ✨ feat

[x] 🐛 fix

[ ] ♻️ refactor

[ ] 💄 style

[ ] 🔨 chore

[ ] 📝 docs

Relevant Issues
N/A (No issue exists for this typo fix)

What is in this change?
This PR corrects a minor typo in an output message within a script.

Before:
echo "Copying to Sever Public"

After:
echo "Copying to Server Public"

Additional Information
This is a minor correction to improve the clarity and professionalism of the script's output. No functional changes are introduced.

Developer Validations
[ ] I ran yarn lint from the root of the repo & committed changes

[ ] Relevant documentation has been updated

[ ] I have tested my code functionality

[ ] Docker build succeeds locally